% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemLabels.R
\name{arch_chemLabels}
\alias{arch_chemLabels}
\title{Chemical Labels with subcripts.}
\usage{
arch_chemLabels(labels_vector)
}
\arguments{
\item{labels_vector}{a vector containing the chemical names.}
}
\value{
A vector with the labels as character including subscripts to display on graphics.
}
\description{
Add subscripts to the chemical labels. Modified from the original version of J. Buxeda i Garriges.
}
\examples{
colnames_df_chem <- c("Fe2O3","SiO2", "Al2O3")
arch_chemLabels(colnames_df_chem)
}
