% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arch_classify.R
\name{arch_classify}
\alias{arch_classify}
\title{Classify by signaries.}
\usage{
arch_classify(dendroMVC)
}
\arguments{
\item{dendroMVC}{Dendrogram of compositional variation matrix (see \code{arch_evennes} function)}
}
\value{
A list of dataframes with \code{output}, \code{signary}, \code{f1}, \code{f2}, \code{m}, \code{des} and \code{zeros} (values that are used in arch_evennes)
}
\description{
Creates signary matrix to obtain balances to plot in a dendrogram in a evenness analysis of the variables
}
