% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropia.R
\name{arch_entropy}
\alias{arch_entropy}
\title{Entropy of the dataset.}
\usage{
arch_entropy(df_chem)
}
\arguments{
\item{df_chem}{a dataframe only with chemical data}
}
\value{
Provides two dataframes based on \code{df_chem}: Entropy and Probability.
}
\description{
Provides the contributions of each element that is taken as a divisor in the ALR transformation to the entropy of total information. 
Code developed from the original code from J.Buxeda i Garrigos.
Entropy: It includes the value of the entropy of the information (in base logarithms 2) (H2)
and the relative value that this entropy has with respect to the maximum that can reach for the number of dimensions (elements) that are take into account (H2%).
The entropy is calculated on the τ.i value, once the total variation (vt) has been subtracted.
Probability: provides the probabilities of each of the divisors estimated from the relative frequency 
that each value τ.i represents with respect to the sum of all these values.
}
\references{
Aitchison, J. (1986). The Statistical Analysis of Compositional Data. 
    \url{https://doi.org/10.1007/978-94-009-4109-0}
    Buxeda i Garrigós, J., & Kilikoglou, V. (2003). 
    Total Variation as a Measure of Variability in Chemical Data Sets. Patterns and Process. 
    A Festchrift in Honor to Dr. Edward Sayre, March, 185–198.
}
