% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arch_evenness.R
\name{arch_evenness}
\alias{arch_evenness}
\title{Chemical Evenness of the dataset.}
\usage{
arch_evenness(df_chem)
}
\arguments{
\item{df_chem}{a dataframe only with chemical data}
}
\value{
Provides evenness plot, MVC dendrogram and CoDa Dendrogram based on \code{df_chem} data.
}
\description{
Provides chemical evenness plot, dendrogram of compositional variation matrix (MVC) and CoDa Dendrogram.
Code developed from the original code from J.Buxeda i Garrigos.
Based on the observations from Aitchison and Buxeda on compositional data.
}
\references{
Aitchison, J. (1986). The Statistical Analysis of Compositional Data.
    \url{https://doi.org/10.1007/978-94-009-4109-0}
    Buxeda i Garrigós, J., & Kilikoglou, V. (2003).
    Total Variation as a Measure of Variability in Chemical Data Sets. Patterns and Process.
    A Festchrift in Honor to Dr. Edward Sayre, March, 185–198.
}
