% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arch_heatmap.R
\name{arch_heatmap}
\alias{arch_heatmap}
\title{Heatmap of the chemical composition.}
\usage{
arch_heatmap(df_chem)
}
\arguments{
\item{df_chem}{chemical data}
}
\value{
A heatmap that is interactive in the html version.
}
\description{
Provides a heatmap showing the chemical outlying values.
}
