% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arch_latex.R
\name{arch_latex}
\alias{arch_latex}
\title{LaTex Formatted Variation Matrix. 
Provides the compositions of previously defined compositional groups in LaTex format ready to create a pdf.}
\usage{
arch_latex(group_list, landscape = FALSE, MVC = FALSE, rm_vars = rm_vars)
}
\arguments{
\item{group_list}{A list of dataframes with each compositional group.}

\item{landscape}{turn TRUE to create a landscape mode table.}

\item{MVC}{TRUE to include matrix of compositional variation}

\item{rm_vars}{include variables that are not to be included in the table.}
}
\value{
LaTex format tables with the compositions of each compositional group.
}
\description{
LaTex Formatted Variation Matrix. 
Provides the compositions of previously defined compositional groups in LaTex format ready to create a pdf.
}
