% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arch_mahala.R
\name{arch_mahala}
\alias{arch_mahala}
\title{Hotteling test based on Mahalanobis distances.}
\usage{
arch_mahala(group_list, test_sample, rm_vars = rm_vars)
}
\arguments{
\item{group_list}{List of dataframes corresponding to known compositional groups.}

\item{test_sample}{A individual which is to be tested}

\item{rm_vars}{Variables that are to be excluded in the analysis.}
}
\value{
Performs Hotteling test based on Mahalanobis distances. Probabilities of memebership to a certain group.
}
\description{
Performs Hotteling test based on Mahalanobis distances to asses to memebership probabilities. 
The variables should be always n+2, where n is the number of individuals of the smallest compositional group.
}
