% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arch_PCA.R
\name{arch_PCA}
\alias{arch_PCA}
\title{Principal Component Analysis.}
\usage{
arch_PCA(
  df_chem,
  df_raw,
  nplot = nplot,
  alr = TRUE,
  printPCA = FALSE,
  labels = FALSE,
  shape_cat_number = 15,
  frame = TRUE,
  PCx = 1,
  PCy = 2,
  nshapes = 10,
  label.size = 3
)
}
\arguments{
\item{df_chem}{chemical data}

\item{df_raw}{categorical data and chemical data}

\item{nplot}{number of plots to display}

\item{alr}{when TRUE applies additive log ratio transformations (default = TRUE)}

\item{printPCA}{when TRUE creates a pdf file with the PCA}

\item{labels}{turn TRUE to display labels}

\item{shape_cat_number}{from 1 to 15, the number of categories to display with}

\item{frame}{= TRUE,}

\item{PCx}{= 1,}

\item{PCy}{=2}

\item{nshapes}{Chose from 1-25 shapes to display all equally, default= 10.}

\item{label.size}{= 3}
}
\value{
A principal component analysis plot.
}
\description{
Plots Principal Component Analysis of alr transformed data.
}
