% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arch_triangles.R
\name{arch_triangles}
\alias{arch_triangles}
\title{Ternary Diagrams.}
\usage{
arch_triangles(
  df_raw,
  plot.category = 1,
  language = 3,
  ordenar = 2,
  c.palette = 2,
  legend = plot.category,
  rounded_circle = 1
)
}
\arguments{
\item{df_raw}{Dataframe with the chemical composition and the catagorical values}

\item{plot.category}{The column with categories to be colored (1,2,3...)}

\item{language}{1 = Catala, 2 = Spanish, 3 = English}

\item{ordenar}{1 = to avoid ordering, 2 = order the dataframe according to the given factor (plot.category)}

\item{c.palette}{0 = black and white, 1 = default, 2 = rainbow, 3= heat.colors, 4 = terrain.colors, 5 = topo.colors, 6 = cm.colors}

\item{legend}{The column from which factors are colored (default = plot.category)}

\item{rounded_circle}{1 = circle around the points, 2 = only dots, without circle}
}
\value{
Diagrams showing different ternary systems based on Fe2O3, MgO, CaO, Al2O3 and SiO2 compositions.
}
\description{
Plots diagrams showing different ternary systems based on Fe2O3, MgO, CaO, Al2O3 and SiO2 compositions.
}
