% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arch_varmat.R
\name{arch_varmat}
\alias{arch_varmat}
\title{Compositional Variation Matrix.
Returns the variation matrix based on the chemical evenness of the dataset. 
Prints the less varying variable to use as divisor in alr transformation
From Buxeda i Garrigós evariation.matrix2 R function}
\usage{
arch_varmat(df_chem)
}
\arguments{
\item{df_chem}{chemical data}
}
\value{
Variation matrix based on \code{df_chem} dataset.
}
\description{
Compositional Variation Matrix.
Returns the variation matrix based on the chemical evenness of the dataset. 
Prints the less varying variable to use as divisor in alr transformation
From Buxeda i Garrigós evariation.matrix2 R function
}
