% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_majolica.R
\docType{data}
\name{df_majolica}
\alias{df_majolica}
\title{Majolica Pottery}
\format{
A \code{dataframe} with 241 observations and 43 variables
\describe{
   \item{SITE}{}
   \item{MUNICIPALLITY}{}
   \item{REGION}{}
   \item{TYPE}{}
   \item{PERIOD}{}
   \item{CHRONOLOGY}{}
   \item{CHRONO.FROM}{}
   \item{CHRONO.TO}{}
   \item{PROVENANCE.MUNICIPALITY}{}
   \item{PROVENANCE.REGION}{}
   \item{GROUP}{}
   \item{Chemical Composition}{}
}
}
\source{
\url{10.1016/j.jas.2007.04.007}
}
\usage{
df_majolica
}
\description{
Compositional data of Majolica pottery
}
\details{
R/df_majolica.R
}
\author{
JJB
}
\keyword{datasets}
