% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arch_dendro.R
\name{arch_dendro}
\alias{arch_dendro}
\title{Hierarchical Clustering Analysis.}
\usage{
arch_dendro(df_chem, df_raw, nplot, printDendro = TRUE, cex = 0.4)
}
\arguments{
\item{df_chem}{chemical data}

\item{df_raw}{categorical data and chemical data}

\item{nplot}{number of plots to display}

\item{printDendro}{when TRUE creates a pdf file with the dendrogram}

\item{cex}{font size}
}
\value{
A dendrogram including \code{df_chem} and \code{df_raw} data.
}
\description{
Plots hierarchical clustering of clr transformed data using centroid algorithm and squared euclidean distance.
}
